Convert nx and ny from Double to Int, and remove one TODO item from README.

+++ app/controllers/Application.scala
@@ -58,7 +58,7 @@ class Grid(val size: Double, val nxOverTwo: Int, val maxIter: Int, val mag: Int,
   val rows = {
     var dxTimesTwo = 2.0 / nxOverTwo
     var dy = dxTimesTwo * sqrt(3) / 2
-    val ny = floor(2.0 / dy)
+    val ny = floor(2.0 / dy).toInt
     var iy = -ny
     dxTimesTwo /= pow(2, mag).toDouble
     val dx = dxTimesTwo / 2
@@ -69,7 +69,7 @@ class Grid(val size: Double, val nxOverTwo: Int, val maxIter: Int, val mag: Int,
       val row: ArrayBuffer[Complex] = ArrayBuffer()
       val isEven = iy % 2 == 0
       var y = iy * dy
-      var nx = floor(sqrt(4.0 / pow(2, 2 * mag).toDouble - y * y) / dx)
+      var nx = floor(sqrt(4.0 / pow(2, 2 * mag).toDouble - y * y) / dx).toInt
       y += c.y
       if ((nx % 2 == 0) != isEven) nx -= 1
       var ix = -nx
